requirements.txt
Flask
openai
python-dotenv
pdfplumber
python-docx
scikit-learn
sentence-transformers